Refactor tests: separate TrailingNewlineRule tests into their own file and clean up BaseRule tests
